Colin fixed some things for me

+++ JIGShape.h
@@ -22,8 +22,8 @@ namespace JIG{
         void virtual submitForRender(void) = 0;
         void virtual updateGeometry(int horizontalVelocity, int verticalVelocity, float rotation) = 0;
         void virtual updateGeometry(int horizontalVelocity, int verticalVelocity, float rotation, KFAColorRGBA color) = 0;
-		int virtual width() = 0;
-		int virtual height() = 0;
+		float virtual width() = 0;
+		float virtual height() = 0;
 		float angle();
 		KFAPoint centerPoint();
 		int getVertexArrayID();
